cart bugfix

+++ src/main/java/com/codecool/shop/dao/implementation/CartDaoMem.java
@@ -1,6 +1,7 @@
 package com.codecool.shop.dao.implementation;
 
 import com.codecool.shop.controller.DbConnect;
+import com.codecool.shop.controller.Util;
 import com.codecool.shop.dao.CartDao;
 import com.codecool.shop.model.Product;
 
@@ -24,15 +25,20 @@ public class CartDaoMem implements CartDao {
 
     @Override
     public void add(Product product) {
-        if (getAll().size() < 10){
-            String query = "INSERT INTO cart (user_id, prod_id, cart_is_active) VALUES (?, ?, true)";
-            try(Connection conn = dataSource.getConnection();
-                PreparedStatement statement = conn.prepareStatement(query)){
-                statement.setInt(1, 1);
-                statement.setInt(2, product.getId());
-                statement.execute();
-            } catch (Exception e){
-                System.out.println(e);
+        if (getSize() < 10){
+            int quantity = Util.productQuantityInCart(product.getId(), dataSource);
+            if(quantity == 0){
+                String query = "INSERT INTO cart (user_id, prod_id, prod_quantity) VALUES (?, ?, 1)";
+                try(Connection conn = dataSource.getConnection();
+                    PreparedStatement statement = conn.prepareStatement(query)){
+                    statement.setInt(1, 1);
+                    statement.setInt(2, product.getId());
+                    statement.execute();
+                } catch (Exception e){
+                    System.out.println(e);
+                }
+            } else {
+                update(product, true);
             }
         } else {
             System.out.println("Your cart is Full!");
@@ -41,12 +47,35 @@ public class CartDaoMem implements CartDao {
 
     @Override
     public void remove(Product product) {
-        String query = "DELETE FROM cart WHERE user_id = 1 AND prod_id = ?;";
+        int quantity = Util.productQuantityInCart(product.getId(), dataSource);
+        if (quantity <= 1){
+            String query = "DELETE FROM cart WHERE user_id = 1 AND prod_id = ?;";
+            try(Connection conn = dataSource.getConnection();
+                PreparedStatement statement = conn.prepareStatement(query)){
+                statement.setInt(1, product.getId());
+                statement.execute();
+            } catch (Exception e) {
+                System.out.println(e);
+            }
+        } else {
+            update(product, false);
+        }
+    }
+
+    @Override
+    public void update(Product product, boolean plus) {
+        int quantity = Util.productQuantityInCart(product.getId(), dataSource);
+        String query = "UPDATE cart SET prod_quantity = ? WHERE prod_id = ?;";
         try(Connection conn = dataSource.getConnection();
             PreparedStatement statement = conn.prepareStatement(query)){
-            statement.setInt(1, product.getId());
+            if(plus){
+                statement.setInt(1, quantity + 1);
+            } else {
+                statement.setInt(1, quantity - 1);
+            }
+            statement.setInt(2, product.getId());
             statement.execute();
-        } catch (Exception e) {
+        } catch (Exception e){
             System.out.println(e);
         }
     }
@@ -54,18 +83,13 @@ public class CartDaoMem implements CartDao {
     @Override
     public Map<Product, Integer> getAll() {
         Map<Product, Integer> cartMap = new HashMap<>();
-        String query = "SELECT * FROM cart WHERE cart_is_active = true AND user_id = 1;";
+        String query = "SELECT * FROM cart WHERE user_id = 1;";
         try(Connection conn = dataSource.getConnection();
             PreparedStatement statement = conn.prepareStatement(query);
             ResultSet result = statement.executeQuery()){
-            Map<Integer, Integer> rawMap = new HashMap<>();
             while (result.next()){
-                int prod_id = result.getInt("prod_id");
-                rawMap.merge(prod_id, 1, Integer::sum);
-            }
-            for (int i: rawMap.keySet()){
-                Product product = productDaoMem.find(i);
-                cartMap.put(product, rawMap.get(i));
+                cartMap.put(productDaoMem.find(result.getInt("prod_id")),
+                        result.getInt("prod_quantity"));
             }
         } catch (Exception e) {
             System.out.println(e);
